server/requirements.txt
Flask
python-dotenv
PyJWT
pymysql
Werkzeug
